refactoring: ProdcutManagement 기능 별 코드 분리
