+++ client/src/hooks/index.tsx
@@ -1,3 +1,4 @@
 export { useLogin } from "./useLogin";
 export { useSignup } from "./useSignup";
 export { useMenuSideabar } from "./useMenuSidebar";
+export { useProductData, useProductGrid, useProductManagement } from "./productManagement";
